[45] Increased the rounding of torque ratios by 2 decimal places.

+++ ATSEngineTool/Database/Entities/TorqueRatio.cs
@@ -22,7 +22,7 @@ namespace ATSEngineTool.Database
         /// 
         /// </summary>
         [Column, Required]
-        public decimal Ratio { get; set; }
+        public double Ratio { get; set; }
 
         #region Foreign Keys
